pyproject.toml
[tool.poetry]
name = "huffpost-mlops"
version = "0.1.0"
description = "MLOps repository for HuffPost news category classification"
authors = ["Codex <codex@example.com>"]
readme = "README.md"
packages = [{ include = "huffpost_classifier" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
torch = "^2.1.0"
pytorch-lightning = "^2.1.0"
torchmetrics = "^1.2.0"
transformers = "^4.36.0"
datasets = "^2.17.0"
hydra-core = "^1.3.2"
scikit-learn = "^1.3.0"
mlflow = "^2.9.2"
dvc = "^3.38.1"
matplotlib = "^3.8.0"
onnx = "^1.14.0"
onnxscript = "^0.2.0"
numpy = "^1.26.0"
pandas = "^2.2.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
isort = "^5.13.0"
flake8 = "^6.1.0"
pre-commit = "^3.6.0"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
